Add a ui package and refactor project
