--- mobile/TaxiApp/app/src/main/java/com/example/taxiapp/MainActivity.java
@@ -1,23 +1,0 @@
-package com.example.taxiapp;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.example.taxiapp.home.HomeFragment;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        if (savedInstanceState == null) {
-            getSupportFragmentManager()
-                    .beginTransaction()
-                    .replace(R.id.main_container, new HomeFragment())
-                    .commit();
-        }
-    }
-}
