+++ mobile/TaxiApp/app/src/main/java/com/example/taxiapp/ui/MainActivity.java
@@ -1,0 +1,39 @@
+package com.example.taxiapp.ui;
+
+import android.os.Bundle;
+import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.drawerlayout.widget.DrawerLayout;
+
+import com.example.taxiapp.R;
+import com.example.taxiapp.ui.home.HomeFragment;
+
+public class MainActivity extends AppCompatActivity {
+
+    private DrawerLayout drawerLayout;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        Toolbar toolbar = findViewById(R.id.app_toolbar);
+        setSupportActionBar(toolbar);
+
+        drawerLayout = findViewById(R.id.drawer_layout);
+
+        // Klik na ikonicu otvara drawer
+        findViewById(R.id.menu_icon).setOnClickListener(v ->
+                drawerLayout.openDrawer(android.view.Gravity.END)
+        );
+
+        if (savedInstanceState == null) {
+            getSupportFragmentManager()
+                    .beginTransaction()
+                    .replace(R.id.main_container, new HomeFragment())
+                    .commit();
+        }
+    }
+}
