+++ mobile/TaxiApp/app/src/main/res/layout/view_toolbar.xml
@@ -1,43 +1,53 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.card.MaterialCardView
+<androidx.appcompat.widget.Toolbar
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/app_toolbar"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_margin="0dp"
-    app:cardElevation="8dp"
-    app:cardCornerRadius="0dp"
-    app:cardBackgroundColor="@android:color/white">
+    android:layout_height="?attr/actionBarSize"
+    android:background="@android:color/transparent"
 
-    <RelativeLayout
+    app:contentInsetStart="0dp"
+    app:contentInsetStartWithNavigation="0dp">
+
+    <com.google.android.material.card.MaterialCardView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="8dp"
-        android:elevation="8dp">
+        android:layout_height="match_parent"
+        app:cardElevation="8dp"
+        app:cardCornerRadius="0dp"
+        app:cardBackgroundColor="@android:color/white">
 
-        <ImageView
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:src="@drawable/taxi_logo"
-            android:id="@+id/logo"/>
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:padding="8dp">
 
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerVertical="true"
-            android:layout_marginStart="16dp"
-            android:layout_toRightOf="@id/logo"
-            android:fontFamily="@font/faster_one"
-            android:text="Taxi App"
-            android:textColor="@color/purple"
-            android:textSize="24dp" />
+            <ImageView
+                android:id="@+id/logo"
+                android:layout_width="50dp"
+                android:layout_height="50dp"
+                android:src="@drawable/taxi_logo" />
 
-        <ImageView
-            android:layout_width="45dp"
-            android:layout_height="48dp"
-            android:layout_alignParentEnd="true"
-            android:layout_marginEnd="1dp"
-            android:src="@drawable/baseline_menu_24" />
-    </RelativeLayout>
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_centerVertical="true"
+                android:layout_marginStart="16dp"
+                android:layout_toEndOf="@id/logo"
+                android:fontFamily="@font/faster_one"
+                android:text="Taxi App"
+                android:textColor="@color/purple"
+                android:textSize="24sp" />
 
-</com.google.android.material.card.MaterialCardView>
+            <ImageView
+                android:id="@+id/menu_icon"
+                android:layout_width="45dp"
+                android:layout_height="48dp"
+                android:layout_alignParentEnd="true"
+                android:src="@drawable/baseline_menu_24" />
+
+        </RelativeLayout>
+
+    </com.google.android.material.card.MaterialCardView>
+
+</androidx.appcompat.widget.Toolbar>
